Change dtype to native int (uint32) for data_size, {x,y}_pixels etc

+++ src/nexgen/nxs_write/NXclassWriters.py
@@ -698,9 +698,9 @@ def write_NXdetector_module(
         ("NXdetector_module",),
     )
 
-    nxmodule.create_dataset("data_origin", data=np.array([0, 0]), dtype=np.uint16)
-    nxmodule.create_dataset("data_size", data=image_size, dtype=np.uint16)
-    nxmodule.create_dataset("data_stride", data=np.array([1, 1]), dtype=np.uint16)
+    nxmodule.create_dataset("data_origin", data=np.array([0, 0]), dtype=np.uint32)
+    nxmodule.create_dataset("data_size", data=image_size, dtype=np.uint32)
+    nxmodule.create_dataset("data_stride", data=np.array([1, 1]), dtype=np.uint32)
 
     # Write fast_ and slow_ pixel_direction
     fast_axis = module["fast_axis"]
@@ -825,17 +825,17 @@ def write_NXcollection(
     # Create detectorSpecific group
     grp = nxdetector.require_group("detectorSpecific")
     grp.create_dataset(
-        "x_pixels", data=detector_params.image_size[1], dtype=np.uint16
+        "x_pixels", data=detector_params.image_size[1], dtype=np.uint32
     )  # fast axis
     grp.create_dataset(
-        "y_pixels", data=detector_params.image_size[0], dtype=np.uint16
+        "y_pixels", data=detector_params.image_size[0], dtype=np.uint32
     )  # slow axis
     # Write these non-spec fields as well because of autoPROC:
     grp.create_dataset(
-        "x_pixels_in_detector", data=detector_params.image_size[1], dtype=np.uint16
+        "x_pixels_in_detector", data=detector_params.image_size[1], dtype=np.uint32
     )  # fast axis
     grp.create_dataset(
-        "y_pixels_in_detector", data=detector_params.image_size[0], dtype=np.uint16
+        "y_pixels_in_detector", data=detector_params.image_size[0], dtype=np.uint32
     )  # slow axis
     if collection_mode == "images":
         grp.create_dataset("nimages", data=num_images)
